Chapte6Review/src/chapte6review/ItemForSale.java
package chapte6review;

public class ItemForSale {

    private double price;
    private String name;
    private int quantity;

    public ItemForSale(String name, double price, int quantity) {
        this.name = name;
        setPrice(price);
        setQuantity(quantity);
    }

    public void setPrice(double price) {
        if (price < 0) {
            this.price = 0;
        } else {
            this.price = price;
        }
    }
    
    public void setQuantity(int quantity) {
        if (quantity < 0) {
            this.quantity = 0;
        } else {
            this.quantity = quantity;
        }
    }

    public double getPrice() {
        return price;
    }

    public String getName() {
        return name;
    }

    public int getQuantity() {
        return quantity;
    }
    
    public double getTotalCost()
    {
        return price * quantity;
    }
}
